Разделение JS проекта на модули
